+++ src/app/(routers)/(board)/community/_components/PostFormClient.tsx
@@ -1,11 +1,8 @@
 'use client';
 
-import type { Editor } from '@tiptap/react';
-
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo } from 'react';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
-import { useEditorConfig } from '@/hooks/editor';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { EditorContent } from '@tiptap/react';
 import { useForm } from 'react-hook-form';
@@ -16,13 +13,12 @@ import { DeleteIcon } from '@/components/icon/icons/Delete';
 import { Toolbar } from '@/components/Toolbar';
 import { Spinner } from '@/components/ui/spinner';
 
+import { usePostEditor } from '../_hooks/usePostEditor';
+import { usePostImages } from '../_hooks/usePostImages';
 import { DesktopPostHeader } from './DesktopPostHeader';
 import { MobilePostHeader } from './MobilePostHeader';
 
 const TITLE_MAX_LENGTH = 30;
-const IMAGE_LIMIT = 2;
-
-const isEditorFilled = (editor: Editor) => !editor.isEmpty && editor.getText().trim().length > 0;
 
 const postSchema = z.object({
   title: z
@@ -32,7 +28,6 @@ const postSchema = z.object({
 });
 
 type PostFormValues = z.infer<typeof postSchema>;
-type ImageItem = { type: 'existing'; url: string } | { type: 'new'; url: string; file: File };
 
 interface PostFormClientProps {
   mode: 'create' | 'edit';
@@ -53,22 +48,18 @@ export function PostFormClient({
 }: PostFormClientProps) {
   const router = useRouter();
 
-  const [images, setImages] = useState<ImageItem[]>(
-    initialImageUrls.map((url) => ({ type: 'existing', url })),
-  );
-  const imagesRef = useRef(images);
+  const { editor, hasEditorContent, contentText, charCountWithoutSpaces } = usePostEditor({
+    initialContent: initialValues?.content,
+  });
 
-  useEffect(() => {
-    imagesRef.current = images;
-  }, [images]);
-
-  useEffect(() => {
-    return () => {
-      imagesRef.current.forEach((item) => {
-        if (item.type === 'new') URL.revokeObjectURL(item.url);
-      });
-    };
-  }, []);
+  const {
+    images,
+    IMAGE_LIMIT,
+    handleImageSelected,
+    handleImageRemove,
+    handleImageSizeExceeded,
+    handleImageLimitExceeded,
+  } = usePostImages(initialImageUrls);
 
   const {
     register,
@@ -83,32 +74,6 @@ export function PostFormClient({
 
   const titleValue = watch('title');
 
-  const editor = useEditorConfig({
-    content: initialValues?.content,
-    variant: 'post',
-    placeholder: '이 곳을 통해 내용을 작성해주세요',
-  });
-
-  const [hasEditorContent, setHasEditorContent] = useState(false);
-  const [contentText, setContentText] = useState('');
-
-  useEffect(() => {
-    if (!editor) return;
-    setHasEditorContent(isEditorFilled(editor));
-    setContentText(editor.getText());
-
-    const onUpdate = () => {
-      setHasEditorContent(isEditorFilled(editor));
-      setContentText(editor.getText());
-    };
-    editor.on('update', onUpdate);
-    return () => {
-      editor.off('update', onUpdate);
-    };
-  }, [editor]);
-
-  const charCountWithoutSpaces = contentText.replace(/\s/g, '').length;
-
   const isSubmitDisabled = !titleValue.trim() || !editor || !hasEditorContent || isSubmitting;
   const headerTitle = mode === 'create' ? '게시물 작성하기' : '게시물 수정하기';
   const submitLabel = mode === 'create' ? '등록' : '수정';
@@ -118,11 +83,11 @@ export function PostFormClient({
     try {
       const contentJson = JSON.stringify(editor.getJSON());
       const newFiles = images
-        .filter((item): item is Extract<ImageItem, { type: 'new' }> => item.type === 'new')
+        .filter((item): item is Extract<typeof item, { type: 'new' }> => item.type === 'new')
         .map((item) => item.file);
       const existingUrls = images
         .filter(
-          (item): item is Extract<ImageItem, { type: 'existing' }> => item.type === 'existing',
+          (item): item is Extract<typeof item, { type: 'existing' }> => item.type === 'existing',
         )
         .map((item) => item.url);
 
@@ -132,34 +97,6 @@ export function PostFormClient({
       toast.error(error instanceof Error ? error.message : '게시물 저장에 실패했습니다.');
     }
   });
-
-  const handleImageSizeExceeded = useCallback(() => {
-    toast.error('이미지 크기는 10MB를 초과할 수 없습니다.');
-  }, []);
-
-  const handleImageLimitExceeded = useCallback(() => {
-    toast.error(`이미지는 최대 ${IMAGE_LIMIT}개까지 첨부할 수 있습니다.`);
-  }, []);
-
-  const handleImageSelected = useCallback(
-    (file: File) => {
-      if (images.length >= IMAGE_LIMIT) {
-        handleImageLimitExceeded();
-        return;
-      }
-      const url = URL.createObjectURL(file);
-      setImages((prev) => [...prev, { type: 'new', url, file }]);
-    },
-    [images.length, handleImageLimitExceeded],
-  );
-
-  const handleImageRemove = (index: number) => {
-    setImages((prev) => {
-      const item = prev[index];
-      if (item.type === 'new') URL.revokeObjectURL(item.url);
-      return prev.filter((_, i) => i !== index);
-    });
-  };
 
   const toolbar = useMemo(
     () => (
